+++ .cruft.json
@@ -1,6 +1,6 @@
 {
   "template": "https://github.com/viseshrp/yapc",
-  "commit": "7e735f29908dbe256f42fe788aa5ad31e905e4cf",
+  "commit": "46945baf7659d518b775239d32b9a66d53964e2c",
   "checkout": null,
   "context": {
     "cookiecutter": {
@@ -16,7 +16,7 @@
       "git_init": "y",
       "github_actions": "y",
       "_template": "https://github.com/viseshrp/yapc",
-      "_commit": "7e735f29908dbe256f42fe788aa5ad31e905e4cf"
+      "_commit": "46945baf7659d518b775239d32b9a66d53964e2c"
     }
   },
   "directory": null,
