+++ pyproject.toml
@@ -46,21 +46,27 @@ wo = "workedon.__main__:main"
 
 [dependency-groups]
 dev = [
-    "pre-commit>=4.2.0",
-    "uv>=0.6.17",
-    "tox>=4.25.0",
-    "tox-uv>=1.25.0",
-    "tox-gh-actions>=3.3.0",
-    "mypy>=1.15.0",
-    "deptry>=0.23.0",
-    "pytest>=8.3.5",
-    "pytest-cov>=6.1.1",
-    "vulture>=2.14",
-    "cogapp>=3.4.1",
-    "twine>=6.1.0",
-    "ipdb>=0.13.13",
-    "hatch>=1.14.1",
-    "freezegun>=1.5.2",
+    "pre-commit",
+    "pre-commit-hooks",  # for check-case-conflict, etc.
+    "uv",
+    "tox",
+    "tox-uv",
+    "tox-gh-actions",
+    "mypy",
+    "deptry",
+    "black",
+    "ruff",
+    "pip-audit",
+    "bandit[toml]",
+    "codespell",
+    "vulture",
+    "pytest",
+    "pytest-cov",
+    "vulture",
+    "cogapp",
+    "twine",
+    "ipdb",
+    "hatch",
 ]
 
 [build-system]
